feat: Implement credit card tracking and expense installment features
This commit introduces new functionality for managing credit cards and handling expense installments.

Changes include:
- Addition of `CardRecord`, `CardStatementResponse`, and `ExpenseInstallmentRecord` DTOs.
- Updates to existing DTOs (`ExpenseRecord`, `RegisterDTO`, `UserRecord`, `WalletRecord`) for integration.
- New database migration script `V10__Credit_Card_Tracking_V1.sql`.
- Modifications to `Card`, `Expense`, `ExpenseInstallment`, `MonthlyExpensesResponse`, `User`, and `Wallet`
entities.
- Introduction of `CardBillingCycleCalculator` and `CardStatementService`.
- Updates to `CardService` and `ExpenseInstallmentService`.
- Changes in `ResourceExceptionHandler`, `SecurityConfiguration`, `SecurityFilter`, and various `Resource` and
`Repository` files to support the new features.
 - Modifications to `application.properties` and `FinanxApplicationTests.java`.

+++ src/main/java/com/example/finanx/DTO/RegisterDTO.java
@@ -2,6 +2,8 @@ package com.example.finanx.DTO;
 
 import com.example.finanx.Entities.UserRole;
 
-public record RegisterDTO(String name, String lastName, Double monthLimit, String email, String password, UserRole role) {
+import java.math.BigDecimal;
+
+public record RegisterDTO(String name, String lastName, BigDecimal monthLimit, String email, String password, UserRole role) {
 
 }
