+++ src/main/java/com/example/finanx/DTO/UserRecord.java
@@ -2,7 +2,9 @@ package com.example.finanx.DTO;
 
 import com.example.finanx.Entities.User;
 
-public record UserRecord(String name, String lastName, Double monthLimit, String email) {
+import java.math.BigDecimal;
+
+public record UserRecord(String name, String lastName, BigDecimal monthLimit, String email) {
 
     public UserRecord(User user) {
         this(user.getName(), user.getLastName(), user.getMonthLimit(), user.getEmail());
